@@ -9,17 +9,17 @@ android {
     vectorDrawables.useSupportLibrary = true
   }
 
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_7
-    targetCompatibility JavaVersion.VERSION_1_7
-  }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 
   resourcePrefix 'swirl_'
 }
 
 dependencies {
-  implementation "com.android.support:support-annotations:${versions.supportLibrary}"
-  implementation "com.android.support:animated-vector-drawable:${versions.supportLibrary}"
+  implementation "androidx.annotation:annotation:${versions.annotation_version}"
+  implementation "androidx.vectordrawable:vectordrawable-animated:${versions.vectordrawable_version}"
 }
 
 apply from: 'gradle-mvn-push.gradle'
